Fix the problem of packaging and EntityService
