+++ src/main/java/com/tunaweza/law/order/LawOrderApplication.java
@@ -3,10 +3,10 @@ package com.tunaweza.law.order;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
+//import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan ({"law-order"})
+//@ComponentScan ({"law-order"})
 public class LawOrderApplication implements CommandLineRunner{
 
 	public static void main(String[] args) {
