+++ src/main/java/com/tunaweza/law/order/controller/SpecialisationController.java
@@ -4,7 +4,9 @@
  */
 package com.tunaweza.law.order.controller;
 
+import com.tunaweza.law.order.repository.SpecialisationRepository;
 import com.tunaweza.law.order.model.Specialisation;
+import com.tunaweza.law.order.service.EntityService;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -15,8 +17,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import service.SpecialisationService;
+import com.tunaweza.law.order.service.SpecialisationService;
 
 /**
  *
@@ -24,17 +27,26 @@ import service.SpecialisationService;
  */
 
 @RestController
+@RequestMapping("/specialisation")
 @Getter
 @Setter
 @AllArgsConstructor
 public class SpecialisationController {
-    @Autowired
-    private final SpecialisationService specialisationService;
+    //
+    private final EntityService specialisationService;
+    
+    @GetMapping("/")
+    public String index(){
+        return "Je suis la page Acceuil";
+    }
+    
     
     @PostMapping("/save")
     public Specialisation add(@RequestBody Specialisation specialisation){
-        return (Specialisation) specialisationService.create(specialisation);
+        specialisationService.create(specialisation);
+        return specialisation;
     }
+    
     
     @GetMapping("/read")
     public List<Specialisation> read(){
@@ -45,4 +57,5 @@ public class SpecialisationController {
     public Specialisation delete(@PathVariable int id){
         return (Specialisation) specialisationService.delete(id);
     }
+
 }
