+++ src/main/java/com/tunaweza/law/order/model/Accuser.java
@@ -13,11 +13,11 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.util.Date;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
-import service.EntityModel;
+import com.tunaweza.law.order.service.EntityModel;
+import lombok.NoArgsConstructor;
 
 /**
  *
@@ -27,7 +27,7 @@ import service.EntityModel;
 @Table(name="Accuser")
 @Getter
 @Setter
-@AllArgsConstructor
+@NoArgsConstructor
 public class Accuser implements EntityModel {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
