+++ src/main/java/com/tunaweza/law/order/model/Culprit.java
@@ -10,8 +10,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import java.util.Date;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
@@ -22,7 +22,7 @@ import lombok.Setter;
 @Table(name="Culprit")
 @Getter
 @Setter
-@AllArgsConstructor
+@NoArgsConstructor
 public class Culprit {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
