+++ src/main/java/com/tunaweza/law/order/repository/AccuserRepository.java
@@ -2,15 +2,15 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package Repository;
+package com.tunaweza.law.order.repository;
 
-import com.tunaweza.law.order.model.Judge;
+import com.tunaweza.law.order.model.Accuser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  *
  * @author azaria
  */
-public interface JudgeRepository extends JpaRepository<Judge,Long> {
+public interface AccuserRepository extends JpaRepository<Accuser,Integer> {
     
 }
