+++ src/main/java/com/tunaweza/law/order/repository/BurdenFileRepository.java
@@ -2,17 +2,15 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package Repository;
+package com.tunaweza.law.order.repository;
 
-import com.tunaweza.law.order.model.Specialisation;
+import com.tunaweza.law.order.model.BurdenFile;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 /**
  *
  * @author azaria
  */
-@Repository
-public interface SpecialisationRepository extends JpaRepository<Specialisation, Integer> {
+public interface BurdenFileRepository extends JpaRepository<BurdenFile,Integer> {
     
 }
