+++ src/main/java/com/tunaweza/law/order/repository/CulpritDefenderRepository.java
@@ -2,15 +2,15 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package Repository;
+package com.tunaweza.law.order.repository;
 
-import com.tunaweza.law.order.model.Culprit;
+import com.tunaweza.law.order.model.CulpritDefender;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  *
  * @author azaria
  */
-public interface CulpritRepository extends JpaRepository<Culprit, Integer> {
+public interface CulpritDefenderRepository extends JpaRepository<CulpritDefender, Integer> {
     
 }
