+++ src/main/java/com/tunaweza/law/order/repository/JudgeRepository.java
@@ -2,15 +2,15 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package Repository;
+package com.tunaweza.law.order.repository;
 
-import com.tunaweza.law.order.model.BurdenFile;
+import com.tunaweza.law.order.model.Judge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  *
  * @author azaria
  */
-public interface BurdenFileRepository extends JpaRepository<BurdenFile,Integer> {
+public interface JudgeRepository extends JpaRepository<Judge,Long> {
     
 }
