+++ src/main/java/com/tunaweza/law/order/service/EntityModel.java
@@ -2,15 +2,12 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package Repository;
-
-import com.tunaweza.law.order.model.Accuser;
-import org.springframework.data.jpa.repository.JpaRepository;
+package com.tunaweza.law.order.service;
 
 /**
  *
  * @author azaria
  */
-public interface AccuserRepository extends JpaRepository<Accuser,Integer> {
+public interface EntityModel {
     
 }
