+++ src/main/java/com/tunaweza/law/order/service/EntityService.java
@@ -2,11 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package service;
+package com.tunaweza.law.order.service;
 
 import com.tunaweza.law.order.model.Specialisation;
 import java.util.List;
-import org.springframework.stereotype.Component;
 
 /**
  *
