--- src/main/java/service/EntityModel.java
@@ -1,13 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
-package service;
-
-/**
- *
- * @author azaria
- */
-public interface EntityModel {
-    
-}
